# docs/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Aire CRM API
  version: 1.0.0
  description: Contrato compartido por la aplicación web y Android nativo.
servers:
  - url: /api/v1
security:
  - firebaseBearer: []
paths:
...
  /billing/tango-invoices:
    get:
      operationId: listTangoInvoices
      tags: [Facturacion]
      parameters:
        - name: company
          in: query
          required: true
          schema: { $ref: '#/components/schemas/TangoCompanyId' }
        - name: fromDate
          in: query
          schema: { type: string, format: date }
        - name: toDate
          in: query
          schema: { type: string, format: date }
        - name: client
          in: query
          schema: { type: string }
        - name: seller
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Comprobantes de Tango filtrados para la compania indicada.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { $ref: '#/components/schemas/TangoInvoiceResult' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /billing/tango-client-invoices:
    get:
      operationId: listClientTangoInvoices
      tags: [Facturacion]
      parameters:
        - name: aireClientId
          in: query
          schema: { type: string }
        - name: srlClientId
          in: query
          schema: { type: string }
        - name: sasClientId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Historial de comprobantes de Tango para un cliente CRM mapeado.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/TangoInvoice' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    firebaseBearer:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
...
  schemas:
...
    TangoCompanyId:
      type: string
      enum: ['4', '5', '6']
      description: '4: Aire/Avion, 5: SRL, 6: SAS.'
    TangoInvoice:
      type: object
      properties:
        FECHA_DE_EMISION: { type: string, format: date-time }
        TIPO_COMPROBANTE: { type: string }
        NRO_COMPROBANTE: { type: string }
        COD_VENDEDOR: { type: string }
        NOMBRE_VENDEDOR: { type: string }
        COD_CLIENTE: { type: string }
        RAZON_SOCIAL: { type: string }
        ID_GVA14: { type: [number, 'null'] }
        TOTAL: { type: [number, 'null'] }
        ID_GVA12: { type: [number, 'null'] }
        ID_GVA23: { type: [number, 'null'] }
        ID_GVA38: { type: [number, 'null'] }
        _company: { type: string }
        _companyId: { $ref: '#/components/schemas/TangoCompanyId' }
    TangoInvoiceResult:
      type: object
      required: [list, sourceTotalCount, filteredCount, truncated]
      properties:
        list:
          type: array
          items: { $ref: '#/components/schemas/TangoInvoice' }
        sourceTotalCount: { type: integer }
        filteredCount: { type: integer }
        truncated: { type: boolean }
...
    Error:
      type: object
      required: [error, code]
      properties:
        error: { type: string }
        code: { type: string }
  responses:
...
    Forbidden:
      description: El usuario no tiene la capacidad requerida.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ValidationError:
      description: La solicitud no cumple el contrato.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }